+++ app/src/main/java/com/jemiola/moodtimeline/views/timeline/TimelienModel.kt
@@ -1,0 +1,6 @@
+package com.jemiola.moodtimeline.views.timeline
+
+interface TimelienModel {
+
+
+}
