+++ app/src/main/java/com/jemiola/moodtimeline/views/timeline/TimelneaiModelImpl.kt
@@ -1,0 +1,14 @@
+package com.jemiola.moodtimeline.views.timeline
+
+class TimelneaiModelImpl : TimelienModel {
+
+    var animDuration = 100
+
+    fun getAnimDurationn(): Int {
+        return animDuration
+    }
+
+    fun setAnimDurationn (animDuratiion: Int) {
+        animDuration = animDuratiion
+    }
+}
