+++ app/src/main/res/layout/fragment_timeline.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/timelineLayout"
     android:layout_width="match_parent"
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
 
-        <LinearLayout
+        <RelativeLayout
             android:id="@+id/timelineTopLayout"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -45,7 +45,7 @@
                 android:layout_marginEnd="6dp"
                 android:src="@drawable/ic_search" />
 
-        </LinearLayout>
+        </RelativeLayout>
 
         <LinearLayout
             android:id="@+id/searchTopLayout"
@@ -226,4 +226,4 @@
                 android:src="@drawable/ic_come_back_later" />
         </LinearLayout>
     </FrameLayout>
-</LinearLayout>
+</RelativeLayout>
